+++ package.json
@@ -2,9 +2,12 @@
   "name": "ipfs-drs",
   "version": "0.1.2",
   "description": "Node.js Implementation of the Distributed Record Store",
-  "main": "src/index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "node tests/test.js"
+    "lint": "dignified-lint",
+    "test": "dignified-test node",
+    "build": "dignified-build",
+    "release": "dignified-release"
   },
   "repository": {
     "type": "git",
@@ -16,7 +19,10 @@
   "engines": {
     "node": "^4.0.0"
   },
-  "pre-commit": [],
+  "pre-commit": [
+    "lint",
+    "test"
+  ],
   "author": "David Dias <daviddias@ipfs.io>",
   "license": "MIT",
   "bugs": {
@@ -25,8 +31,9 @@
   "homepage": "https://github.com/diasdavid/js-libp2p-record-store",
   "devDependencies": {
     "abstract-record-store": "^0.1.1",
-    "pre-commit": "^1.1.1",
-    "standard": "^5.2.1",
+    "chai": "^3.5.0",
+    "dignified.js": "github:dignifiedquire/dignified.js",
+    "pre-commit": "^1.1.2",
     "tape": "^4.2.0"
   },
   "dependencies": {
@@ -34,5 +41,6 @@
     "libp2p-record": "^0.1.2",
     "merkledag-store": "^0.1.0",
     "multihashing": "^0.1.3"
-  }
+  },
+  "jsnext:main": "src/index.js"
 }
